+++ ElvUI_EltreumUI/Core/Options.lua
@@ -3819,46 +3819,26 @@ The Item Level shown on the Character Panel Skin uses code from Simple Item leve
 	ElvUI_EltreumUI.Options.args.changelog.args.added = ACH:Group(E:TextGradient("Added", 0.50, 0.70, 1, 0.67, 0.95, 1), nil, 3)
 	ElvUI_EltreumUI.Options.args.changelog.args.added.inline = true
 	ElvUI_EltreumUI.Options.args.changelog.args.added.args.description = ACH:Description([[
-Added a test function, /eltruism encountercheck, which will block some functions from running during Encounters
-Added several NPC ids for the classification icons tag
-Added more models to the rotation fix for 3D portraits
-Added options to reverse Focus gradient direction
-Added a reload prompt when exiting Edit Mode
-Added Tank and Assist unitframe borders
-Added BigWigs nameplate icons settings
+Added /eltruism togglemodels command to disable 3D models on unitframes, this is recommended due to a bug on Blizzard's end with 3D models causing a big performance impact
+Added a new button in Eltruism > Aura Filters to more easily change filters on all frames
+Added movers to AuctionHouseFrame, TradeSkillFrame, CraftFrame and ProfessionsFrame
 ]], 3, "small", nil, nil, nil, nil, "full")
 	ElvUI_EltreumUI.Options.args.changelog.args.updated = ACH:Group(E:TextGradient("Updated", 0.50, 0.70, 1, 0.67, 0.95, 1), nil, 3)
 	ElvUI_EltreumUI.Options.args.changelog.args.updated.inline = true
 	ElvUI_EltreumUI.Options.args.changelog.args.updated.args.description = ACH:Description([[
-Updated Objective Tracker Skin to not skin the Quest Item, this can still cause issues and is likely to end up hiding the Quest Item if none can be found
-Updated /eltruism paging to accept arguments such as /eltruism paging 3 to page a specific action bar
-Updated [eltruism:stance] to once again display stances in Retail, such as paladin auras
-Updated Class Combat Icons to attempt to detect if an enemy is melee or ranged
-Updated action paging to properly display in some World Quests
-Updated LootText to only replace font when it is enabled
-Updated Blizzard's Edit mode import function
-Updated German locale by Dlarge
-Updated for Classic Era 1.15.4
+Updated shadows to better handle adding shadows when Borders are enabled
+Updated Aura Filters, its now recommended to use the Minimal Filter
+Updated profiles to now use the Minimal Aura Filters during install
+Updated profile to disable ElvUI Abbreviating Enchants
+Updated profiles to now disable Models during install
+Updated some of the version checking
 ]], 5, "small", nil, nil, nil, nil, "full")
 	ElvUI_EltreumUI.Options.args.changelog.args.fixed = ACH:Group(E:TextGradient("Fixed", 0.50, 0.70, 1, 0.67, 0.95, 1), nil, 4)
 	ElvUI_EltreumUI.Options.args.changelog.args.fixed.inline = true
 	ElvUI_EltreumUI.Options.args.changelog.args.fixed.args.description = ACH:Description([[
-Fixed an issue where the Objective Tracker was not moving with the ElvUI Mover due to it being locked into the display zone
-Fixed Addon list not being saved with the debug commands due to a change in the API
-Fixed an issue by removing the Battleground Map mover since it can be moved in game
-Fixed a possible issue with the Talent Scale when it was displayed in combat
-Fixed Arena unitframes missing textures when custom textures was enabled
-Fixed some missing movers disappearing in ElvUI mover mode
-Fixed some of ProjectAzilroka's profile not being imported
-Fixed an error while importing Shadow and Light's profile
-Fixed Target and Focus gradient color reverse not working
-Fixed the Cooldown Pulse appearing in Cinematics
-Fixed a possible error with the Instance Text
-Fixed a compatibility issue with MerathilisUI
-Fixed an issue in Cataclysm with RareScanner
-Fixed Delves objective skin
-Fixed Immersion Skinning
-Fixed Scrap skinning
+Fixed an issue with Borders where some Buffs/Debuffs could have the wrong color on Unitframes
+Fixed /eltruism paging NUMBER command incorrectly setting visibility settings
+Fixed Windtools's GameBar shadows
 ]], 7, "small", nil, nil, nil, nil, "full")
 
 	--[[