+++ Changelog.md
@@ -1,6 +1,19 @@
 ### 🌌 Next Version
 The future is uncertain
 ___
+### 4.2.4
+Added /eltruism togglemodels command to disable 3D models on unitframes, this is recommended due to a bug on Blizzard's end with 3D models causing a big performance impact  
+Added a new button in Eltruism > Aura Filters to more easily change filters on all frames  
+Added movers to AuctionHouseFrame, TradeSkillFrame, CraftFrame and ProfessionsFrame  
+Updated shadows to better handle adding shadows when Borders are enabled  
+Updated Aura Filters, its now recommended to use the Minimal Filter  
+Updated profiles to now use the Minimal Aura Filters during install  
+Updated profile to disable ElvUI Abbreviating Enchants  
+Updated profiles to now disable Models during install  
+Updated some of the version checking  
+Fixed an issue with Borders where some Buffs/Debuffs could have the wrong color on Unitframes  
+Fixed /eltruism paging NUMBER command incorrectly setting visibility settings  
+Fixed Windtools's GameBar shadows  
 ### 4.2.3.1
 Added a test function, /eltruism encountercheck, which will block some functions from running during Encounters  
 Updated [eltruism:stance] to once again display stances in Retail, such as paladin auras